Output wrapper
Whenever the validator produces output, several conditions
have to be checked to make sure that the desired output
is delivered to the user / calling program in the expected way.

This wrapper makes sure that if we're in porcelain mode, any
output will be formatted as Json-Patch. Additionally, it adheres
to the silent-flag.